YALI-44: UX Recommendation: Event Geocoder
Updated Download status when the event count changes to 0 (meaning download is required because events were deleted or don't exist).
Updated Geocoder status when the event count changes to 0.
Updated TabControls to add an orange alert as a reminder to publish the map if it has not yet been published.
Reconfigured TabControls to better suport the publish reminder, errors, and warnings.
Removed extra console log when saving.